read whole class into memory when constructing class reader
This provides massive performance boosts. Also all methods delegate to
their respective counterparts in DataInputStream. A CountingInputStream
from Guava provides the position. This solves the problem for position
reporting for UTF strings.

+++ src/main/java/com/github/musikk/classreader/ClassReader.java
@@ -138,6 +138,6 @@ public interface ClassReader {
 	 *
 	 * @return the current position
 	 */
-	int getPosition();
+	long getPosition();
 
 }
